Created a new helper function header
